# guest: add plugin.Set for simpler example and fixes URL parsing