--- a/examples/nodenumber/main.go
+++ b/examples/nodenumber/main.go
@@ -1,5 +1,3 @@
-//go:build tinygo.wasm
-
 /*
    Copyright 2023 The Kubernetes Authors.
 
@@ -20,28 +18,109 @@
 // '-target=wasi'. See /guest/RATIONALE.md for details.
 package main
 
-// Override the default GC with a more performant one.
-// Note: this requires tinygo flags: -gc=custom -tags=custommalloc
 import (
-	_ "github.com/wasilibs/nottinygc"
+	"encoding/json"
+	"fmt"
 
-	"sigs.k8s.io/kube-scheduler-wasm-extension/examples/nodenumber/plugin"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/config"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/enqueue"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/prescore"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/score"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/plugin"
 )
 
 // main is compiled to a WebAssembly function named "_start", called by the
 // wasm scheduler plugin during initialization.
 func main() {
-	plugin, err := plugin.New(config.Get())
-	if err != nil {
-		panic(err)
+	var args nodeNumberArgs
+	if jsonConfig := config.Get(); jsonConfig != nil {
+		if err := json.Unmarshal(jsonConfig, &args); err != nil {
+			panic(fmt.Errorf("decode arg into NodeNumberArgs: %w", err))
+		}
 	}
-	// Below is like `var _ api.EnqueueExtensions = plugin`, except it also
-	// wires up functions the host should provide (go:wasmimport).
-	enqueue.SetPlugin(plugin)
-	prescore.SetPlugin(plugin)
-	score.SetPlugin(plugin)
+	plugin.Set(&NodeNumber{reverse: args.Reverse})
+}
+
+// NodeNumber is an example plugin that favors nodes that share a numerical
+// suffix with the pod name.
+//
+// For example, when a pod named "Pod1" is scheduled, a node named "Node1" gets
+// a higher score than a node named "Node9".
+//
+// # Notes
+//
+//   - Only the last character in names are considered. This means "Node99" is
+//     treated the same as "Node9"
+//   - The reverse field inverts the score. For example, when `reverse == true`
+//     a numeric match gets a results in a lower score than a match.
+type NodeNumber struct {
+	reverse bool
+}
+
+type nodeNumberArgs struct {
+	Reverse bool `json:"reverse"`
+}
+
+const (
+	// Name is the name of the plugin used in the plugin registry and configurations.
+	Name             = "NodeNumber"
+	preScoreStateKey = "PreScore" + Name
+)
+
+// preScoreState computed at PreScore and used at Score.
+type preScoreState struct {
+	podSuffixNumber uint8
+}
+
+// EventsToRegister implements api.EnqueueExtensions
+func (pl *NodeNumber) EventsToRegister() []api.ClusterEvent {
+	return []api.ClusterEvent{
+		{Resource: api.Node, ActionType: api.Add},
+	}
+}
+
+// PreScore implements api.PreScorePlugin
+func (pl *NodeNumber) PreScore(state api.CycleState, pod proto.Pod, _ proto.NodeList) *api.Status {
+	podnum, ok := lastNumber(pod.Spec().GetNodeName())
+	if !ok {
+		return nil // return success even if its suffix is non-number.
+	}
+
+	state.Write(preScoreStateKey, &preScoreState{podSuffixNumber: podnum})
+	return nil
+}
+
+// Score implements api.ScorePlugin
+func (pl *NodeNumber) Score(state api.CycleState, _ proto.Pod, nodeName string) (int32, *api.Status) {
+	var match bool
+	if data, ok := state.Read(preScoreStateKey); ok {
+		// Match is when there is a last digit, and it is the pod suffix.
+		nodenum, ok := lastNumber(nodeName)
+		match = ok && data.(*preScoreState).podSuffixNumber == nodenum
+	} else {
+		// Match is also when there is no pod spec node name.
+		match = true
+	}
+
+	if pl.reverse {
+		match = !match // invert the condition.
+	}
+
+	if match {
+		return 10, nil
+	}
+	return 0, nil
+}
+
+// lastNumber returns the last number in the string or false.
+func lastNumber(str string) (uint8, bool) {
+	if len(str) == 0 {
+		return 0, false
+	}
+
+	// We have at least a single character name. See if the last is a digit.
+	lastChar := str[len(str)-1]
+	if '0' <= lastChar && lastChar <= '9' {
+		return lastChar - '0', true
+	}
+	return 0, false
 }