--- a/guest/testdata/cyclestate/main.go
+++ b/guest/testdata/cyclestate/main.go
@@ -18,13 +18,9 @@
 // See https://kubernetes.io/docs/concepts/scheduling-eviction/scheduling-framework/#extension-points
 package main
 
-// Override the default GC with a more performant one.
-// Note: this requires tinygo flags: -gc=custom -tags=custommalloc
 import (
 	"os"
 	"unsafe"
-
-	_ "github.com/wasilibs/nottinygc"
 
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"