--- a/guest/testdata/score/main.go
+++ b/guest/testdata/score/main.go
@@ -16,12 +16,8 @@
 
 package main
 
-// Override the default GC with a more performant one.
-// Note: this requires tinygo flags: -gc=custom -tags=custommalloc
 import (
 	"os"
-
-	_ "github.com/wasilibs/nottinygc"
 
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"