@codefromthecrypt 2023-08-02
added polish, mostly README